[AIDAPP-1303]: test - centralize enablePreference helper and drop low-value automated-status tests
