--- app-modules/service-management/tests/Tenant/Models/ServiceRequestTypeTest.php
@@ -1,54 +1,0 @@
-<?php
-
-/*
-<COPYRIGHT>
-
-    Copyright © 2016-2026, Canyon GBS Inc. All rights reserved.
-
-    Aiding App® is licensed under the Elastic License 2.0. For more details,
-    see <https://github.com/canyongbs/aidingapp/blob/main/LICENSE.>
-
-    Notice:
-
-    - You may not provide the software to third parties as a hosted or managed
-      service, where the service provides users with access to any substantial set of
-      the features or functionality of the software.
-    - You may not move, change, disable, or circumvent the license key functionality
-      in the software, and you may not remove or obscure any functionality in the
-      software that is protected by the license key.
-    - You may not alter, remove, or obscure any licensing, copyright, or other notices
-      of the licensor in the software. Any use of the licensor’s trademarks is subject
-      to applicable law.
-    - Canyon GBS Inc. respects the intellectual property rights of others and expects the
-      same in return. Canyon GBS® and Aiding App® are registered trademarks of
-      Canyon GBS Inc., and we are committed to enforcing and protecting our trademarks
-      vigorously.
-    - The software solution, including services, infrastructure, and code, is offered as a
-      Software as a Service (SaaS) by Canyon GBS Inc.
-    - Use of this software implies agreement to the license terms and conditions as stated
-      in the Elastic License 2.0.
-
-    For more information or inquiries please visit our website at
-    <https://www.canyongbs.com> or contact us via email at legal@canyongbs.com.
-
-</COPYRIGHT>
-*/
-
-use AidingApp\ServiceManagement\Models\ServiceRequestStatus;
-use AidingApp\ServiceManagement\Models\ServiceRequestType;
-use App\Features\AutomatedStatusChangeOnAssignmentFeature;
-
-it('persists the automated status change fk and loads the relation', function () {
-    $status = ServiceRequestStatus::factory()->open()->create();
-
-    $type = ServiceRequestType::factory()->create([
-        'automated_status_id' => $status->getKey(),
-    ]);
-
-    expect($type->fresh()->automated_status_id)->toBe($status->getKey())
-        ->and($type->automatedStatus->is($status))->toBeTrue();
-});
-
-it('has the automated status change feature active because the migration activated it', function () {
-    expect(AutomatedStatusChangeOnAssignmentFeature::active())->toBeTrue();
-});
