+++ app-modules/service-management/tests/Tenant/Observers/ServiceRequestObserverTest.php
@@ -44,7 +44,6 @@ use AidingApp\ServiceManagement\Models\ServiceRequest;
 use AidingApp\ServiceManagement\Models\ServiceRequestPriority;
 use AidingApp\ServiceManagement\Models\ServiceRequestStatus;
 use AidingApp\ServiceManagement\Models\ServiceRequestType;
-use AidingApp\ServiceManagement\Models\ServiceRequestTypeEmailPreference;
 use AidingApp\ServiceManagement\Models\ServiceRequestTypeEmailTemplate;
 use AidingApp\ServiceManagement\Notifications\SendClosedServiceFeedbackNotification;
 use AidingApp\ServiceManagement\Notifications\SendEducatableServiceRequestClosedNotification;
@@ -57,24 +56,7 @@ use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Notification;
 
 use function Pest\Laravel\actingAs;
-
-if (! function_exists('enablePreference')) {
-    function enablePreference(
-        ServiceRequestType $type,
-        ServiceRequestEmailTemplateType $templateType,
-        ServiceRequestTypeEmailTemplateRole $role,
-        ServiceRequestNotificationChannel $channel,
-        bool $isEnabled = true,
-    ): void {
-        ServiceRequestTypeEmailPreference::create([
-            'service_request_type_id' => $type->getKey(),
-            'service_request_email_template_type' => $templateType,
-            'service_request_email_template_role' => $role,
-            'notification_channel' => $channel,
-            'is_enabled' => $isEnabled,
-        ]);
-    }
-}
+use function Tests\enablePreference;
 
 describe('Created → Customer', function () {
     it('sends customer created notification when preference is enabled and status is Open', function () {
