+++ app-modules/service-management/tests/Tenant/Observers/ServiceRequestUpdateObserverTest.php
@@ -45,7 +45,6 @@ use AidingApp\ServiceManagement\Models\ServiceRequest;
 use AidingApp\ServiceManagement\Models\ServiceRequestPriority;
 use AidingApp\ServiceManagement\Models\ServiceRequestStatus;
 use AidingApp\ServiceManagement\Models\ServiceRequestType;
-use AidingApp\ServiceManagement\Models\ServiceRequestTypeEmailPreference;
 use AidingApp\ServiceManagement\Models\ServiceRequestTypeEmailTemplate;
 use AidingApp\ServiceManagement\Models\ServiceRequestUpdate;
 use AidingApp\ServiceManagement\Notifications\SendEducatableServiceRequestUpdatedNotification;
@@ -53,23 +52,7 @@ use AidingApp\ServiceManagement\Notifications\ServiceRequestUpdated;
 use App\Models\User;
 use Illuminate\Support\Facades\Notification;
 
-if (! function_exists('enablePreference')) {
-    function enablePreference(
-        ServiceRequestType $type,
-        ServiceRequestEmailTemplateType $templateType,
-        ServiceRequestTypeEmailTemplateRole $role,
-        ServiceRequestNotificationChannel $channel,
-        bool $isEnabled = true,
-    ): void {
-        ServiceRequestTypeEmailPreference::create([
-            'service_request_type_id' => $type->getKey(),
-            'service_request_email_template_type' => $templateType,
-            'service_request_email_template_role' => $role,
-            'notification_channel' => $channel,
-            'is_enabled' => $isEnabled,
-        ]);
-    }
-}
+use function Tests\enablePreference;
 
 if (! function_exists('createServiceRequestWithAssignment')) {
     function createServiceRequestWithAssignment(ServiceRequestType $type, User $assignedUser): ServiceRequest
